Add ligature lookups to the font file (#2)
* write empty latin characters needed for ligatures

All glyphes that should be replaced by a ligatures have to be included
in the font. This is a subset of latin characters we use to create
ligatures from

* create lookup tables and add every glyph

Create a lookup table for latin ligature substitution.
The glyph name is used as ligature so and we only use the
characters [a-zA-Z0-9_] to create ligatures. That means
that only these characters should appear in glyph names
or else the lookup table will not work (i.e. no ligatures at all)